move waitForConfigsFromAD to cmd/agent/common
