+++ cmd/agent/common/autodiscovery.go
@@ -17,6 +17,7 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/config"
 	confad "github.com/DataDog/datadog-agent/pkg/config/autodiscovery"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
+	"go.uber.org/atomic"
 )
 
 // This is due to an AD limitation that does not allow several listeners to work in parallel
@@ -181,6 +182,66 @@ func setupAutoDiscovery(confSearchPaths []string, metaScheduler *scheduler.MetaS
 	}
 
 	return ad
+}
+
+// schedulerFunc is a type alias to allow a function to be used as an AD scheduler
+type schedulerFunc func([]integration.Config)
+
+// Schedule implements scheduler.Scheduler#Schedule.
+func (sf schedulerFunc) Schedule(configs []integration.Config) {
+	sf(configs)
+}
+
+// Unschedule implements scheduler.Scheduler#Unschedule.
+func (sf schedulerFunc) Unschedule(configs []integration.Config) {
+	// (do nothing)
+}
+
+// Stop implements scheduler.Scheduler#Stop.
+func (sf schedulerFunc) Stop() {
+}
+
+// WaitForConfigsFromAD waits until a count of discoveryMinInstances configs
+// with name checkName are scheduled by AD, and returns the matches.  It does
+// so by subscribing to the AD metascheduler.  If the context is cancelled
+// then any accumulated configs are returned, even if that is fewer than
+// discoveryMinInstances.
+func WaitForConfigsFromAD(ctx context.Context, checkName string, discoveryMinInstances int) (configs []integration.Config) {
+	configChan := make(chan integration.Config)
+
+	// signal to the scheduler when we are no longer waiting, so we do not continue
+	// to push items to configChan
+	waiting := atomic.NewBool(true)
+	defer func() {
+		waiting.Store(false)
+		// ..and drain any message currently pending in the channel
+		select {
+		case <-configChan:
+		default:
+		}
+	}()
+
+	// add the scheduler in a goroutine, since it will schedule any "catch-up" immediately,
+	// placing items in configChan
+	go AC.AddScheduler("check-cmd", schedulerFunc(func(configs []integration.Config) {
+		for _, cfg := range configs {
+			if cfg.Name == checkName {
+				if waiting.Load() {
+					configChan <- cfg
+				}
+			}
+		}
+	}), true)
+
+	for len(configs) < discoveryMinInstances {
+		select {
+		case cfg := <-configChan:
+			configs = append(configs, cfg)
+		case <-ctx.Done():
+			return
+		}
+	}
+	return
 }
 
 // WaitForConfigs retries the collection of Autodiscovery configs until the checkMatcher function (which
