+++ cmd/agent/common/commands/check.go
@@ -21,7 +21,6 @@ import (
 
 	"github.com/fatih/color"
 	"github.com/spf13/cobra"
-	"go.uber.org/atomic"
 	"gopkg.in/yaml.v2"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/app/standalone"
@@ -169,8 +168,9 @@ func Check(loggerName config.LoggerName, confFilePath *string, flagNoColor *bool
 			// AutoDiscovery.  NOTE: we do not start common.Coll, either.
 			collector.InitCheckScheduler(common.Coll)
 
-			waitForConfigsFromADCtx, cancelTimeout := context.WithTimeout(context.Background(), time.Duration(discoveryTimeout)*time.Second)
-			allConfigs := waitForConfigsFromAD(waitForConfigsFromADCtx, checkName, int(discoveryMinInstances))
+			waitCtx, cancelTimeout := context.WithTimeout(
+				context.Background(), time.Duration(discoveryTimeout)*time.Second)
+			allConfigs := common.WaitForConfigsFromAD(waitCtx, checkName, int(discoveryMinInstances))
 			cancelTimeout()
 
 			// make sure the checks in cs are not JMX checks
@@ -454,65 +454,6 @@ func Check(loggerName config.LoggerName, confFilePath *string, flagNoColor *bool
 	}
 	setupCmd(cmd)
 	return cmd
-}
-
-// schedulerFunc is a type alias to allow a function to be used as an AD scheduler
-type schedulerFunc func([]integration.Config)
-
-// Schedule implements scheduler.Scheduler#Schedule.
-func (sf schedulerFunc) Schedule(configs []integration.Config) {
-	sf(configs)
-}
-
-// Unschedule implements scheduler.Scheduler#Unschedule.
-func (sf schedulerFunc) Unschedule(configs []integration.Config) {
-	// (do nothing)
-}
-
-// Stop implements scheduler.Scheduler#Stop.
-func (sf schedulerFunc) Stop() {
-}
-
-// waitForConfigsFromAD waits until a count of discoveryMinInstances configs
-// with name checkName are scheduled by AD, and returns the matches.  It does
-// so by subscribing to the AD metascheduler.  If the context is cancelled
-// then any accumulated configs are returned, even if that is fewer than
-// discoveryMinInstances.
-func waitForConfigsFromAD(ctx context.Context, checkName string, discoveryMinInstances int) (configs []integration.Config) {
-	configChan := make(chan integration.Config)
-
-	// signal to the scheduler when we are no longer waiting, so we do not continue
-	// to push items to configChan
-	waiting := atomic.NewBool(true)
-	defer func() {
-		waiting.Store(false)
-		select {
-		case <-configChan:
-		default:
-		}
-	}()
-
-	// add the scheduler in a goroutine, since it will schedule any "catch-up" immediately,
-	// placing items in configChan
-	go common.AC.AddScheduler("check-cmd", schedulerFunc(func(configs []integration.Config) {
-		for _, cfg := range configs {
-			if cfg.Name == checkName {
-				if waiting.Load() {
-					configChan <- cfg
-				}
-			}
-		}
-	}), true)
-
-	for len(configs) < discoveryMinInstances {
-		select {
-		case cfg := <-configChan:
-			configs = append(configs, cfg)
-		case <-ctx.Done():
-			return
-		}
-	}
-	return
 }
 
 func runCheck(c check.Check, demux aggregator.Demultiplexer) *check.Stats {
